refactor: SessionProvider removed since it won't be used. useLocalStorage is only client component hook error fixed.
