+++ package.json
@@ -16,7 +16,7 @@
   },
   "dependencies": {
     "@auth/drizzle-adapter": "^1.1.0",
-    "@nextui-org/react": "^2.4.0",
+    "@nextui-org/react": "^2.4.1",
     "@react-email/components": "^0.0.12",
     "@react-email/render": "^0.0.10",
     "@t3-oss/env-nextjs": "^0.10.1",
