--- src/app/[locale]/(content)/layout.tsx
@@ -1,13 +1,0 @@
-"use client";
-import { Toaster } from "@/src/components/customs/Sonner";
-import { SessionProvider } from "next-auth/react";
-import React from "react";
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return (
-    <SessionProvider>
-      {children}
-      <Toaster />
-    </SessionProvider>
-  );
-}
