+++ src/app/[locale]/(search)/layout.tsx
