+++ src/app/[locale]/(search)/search/loading.tsx
