+++ src/app/[locale]/(search)/search/page.tsx
