+++ src/app/[locale]/dashboard/create-word/page.tsx
