+++ src/app/[locale]/dashboard/layout.tsx
