+++ src/app/[locale]/dashboard/page.tsx
