+++ src/app/[locale]/forgot-password/page.tsx
