+++ src/app/[locale]/reset-password/[id]/page.tsx
