+++ src/lib/uploadthing.ts
@@ -1,7 +1,7 @@
 import { generateComponents } from "@uploadthing/react";
 import { generateReactHelpers } from "@uploadthing/react/hooks";
 
-import type { OurFileRouter } from "@/app/api/uploadthing/core";
+import type { OurFileRouter } from "@/src/app/api/uploadthing/core";
 
 export const { UploadButton, UploadDropzone, Uploader } =
   generateComponents<OurFileRouter>();
